YUK-7839 DynamicPointDispatch table thinks devices are disabled when they aren't Fixed wrong int use. Updated comment.

+++ yukon-server/Database/include/tbl_pt_alarm.h
@@ -137,7 +137,7 @@ public:
    virtual BOOL getNotifyOnAcknowledge()            const;
    virtual BOOL getNotifyOnClear()                  const;
    virtual UINT getNotificationGroupID()            const;
-   UINT         getPointID()                        const;
+   LONG         getPointID()                        const;
 
    virtual bool isNotifyExcluded( int alarm ) const;
    virtual bool isAutoAcked     ( int alarm ) const;
